+++ src/opender/active_power_support_funcs/p_funcs.py
@@ -18,7 +18,7 @@ from opender.active_power_support_funcs import active_power_limit, frequency_dro
 class DesiredActivePower:
     """
     Desired active power calculation from active power support functions
-    EPRI Report Reference: Section 3.7 in Report #3002025583: IEEE 1547-2018 OpenDER Model
+    EPRI Report Reference: Section 3.7 in Report #3002026631: IEEE 1547-2018 OpenDER Model
     """
 
     def __init__(self, der_obj):
@@ -43,7 +43,7 @@ class DesiredActivePower:
         self.freqdroop = frequency_droop.FreqDroop(self.der_obj)
         self.enterserviceperf = es_perf.EnterServicePerformance(der_obj)
 
-    def calculate_p_funcs(self, p_out_pu):
+    def calculate_p_funcs(self, p_out_w):
         """
         Call active power support functions and based on the results, generate desired active power output
 
@@ -58,22 +58,22 @@ class DesiredActivePower:
         self.p_pv_limit_pu = self.voltwatt.calculate_p_pv_limit_pu()
 
         # Frequency-droop function
-        self.p_pf_pu, self.pf_uf_active, self.pf_of_active = self.freqdroop.calculate_p_pf_pu(p_out_pu,
+        self.p_pf_pu, self.pf_uf_active, self.pf_of_active = self.freqdroop.calculate_p_pf_pu(p_out_w,
                                                                                               self.ap_limit_rt,
                                                                                               self.p_pv_limit_pu)
         # Enter service ramp performance
         self.p_es_pu = self.enterserviceperf.es_performance()
 
         # Calculate final desired active power based on other functions
-        self.p_desired_pu = self.calculate_p_desired_pu(p_out_pu)
+        self.p_desired_pu = self.calculate_p_desired_pu(p_out_w)
 
         return self.p_desired_pu
 
-    def calculate_p_desired_pu(self, p_out_pu):
+    def calculate_p_desired_pu(self, p_out_w):
         """
         Based on the calculated values from volt-watt, frequency-droop, active power limit, and enter service ramp,
         their enabling signal, and DER operating status, generate the DER desired active power output
-        EPRI Report Reference: Section 3.7.1.5 in Report #3002025583: IEEE 1547-2018 OpenDER Model
+        EPRI Report Reference: Section 3.7.1.5 in Report #3002026631: IEEE 1547-2018 OpenDER Model
 
         Variable used in this function:
         :param ap_limit_enable_exec:	Active power limit enable (AP_LIMIT_ENABLE) signal after execution delay