+++ src/opender/auxiliary_funcs/flipflop.py
@@ -21,7 +21,7 @@
 class FlipFlop:
     """
     |  Flipflop logic
-    |  EPRI Report Reference: Section 3.12.5 in Report #3002025583: IEEE 1547-2018 OpenDER Model
+    |  EPRI Report Reference: Section 3.12.5 in Report #3002026631: IEEE 1547-2018 OpenDER Model
     """
     def __init__(self, ff_init):
         self.ff_out_prev = ff_init