+++ src/opender/op_cond_proc.py
@@ -28,7 +28,7 @@ class DERInputs:
     """
     This module handles inputs to the DER model, and generate per unit value of applicable voltage,
     per unit available power for PV DER, per unit active power demand for BESS DER, etc.
-    EPRI Report Reference: Section 3.3 in Report #3002025583: IEEE 1547-2018 OpenDER Model
+    EPRI Report Reference: Section 3.3 in Report #3002026631: IEEE 1547-2018 OpenDER Model
     """
 
     def __init__(self, der_file: DERCommonFileFormat):
@@ -70,7 +70,7 @@ class DERInputs:
 
     def operating_condition_input_processing(self):
         """
-        EPRI Report Reference: Section 3.3 in Report #3002025583: IEEE 1547-2018 OpenDER Model
+        EPRI Report Reference: Section 3.3 in Report #3002026631: IEEE 1547-2018 OpenDER Model
 
         Used variables as inputs:
         :param v_a, v_b, v_c:	Three-phase line-to-ground RMS voltage at RPA
@@ -157,7 +157,7 @@ class DERInputs:
     def operating_conditions_validity_check(self):
         """
         Validity Check for DER Model operating conditions
-        Reference: Table 3-5 in Report #3002025583: IEEE 1547-2018 OpenDER Model
+        Reference: Table 3-5 in Report #3002026631: IEEE 1547-2018 OpenDER Model
         Should be executed every timestep
         """
 
@@ -167,6 +167,10 @@ class DERInputs:
             if self.v < 0:
                 logging.error("Error: V should be greater than 0, converting it to postive")
                 self.v = -self.v
+            if self.theta is None:
+                logging.warning("Error: Theta is not defined. Default to 0")
+                self.theta = 0
+
 
         if self.der_file.NP_PHASE == "THREE":
             if self.v_a is None or self.v_b is None or self.v_c is None: