+++ src/opender/reactive_power_support_funcs/constant_pf.py
@@ -23,7 +23,7 @@ from opender.auxiliary_funcs.time_delay import TimeDelay
 class ConstantPowerFactor:
     """
     Constant Power Factor Function
-    EPRI Report Reference: Section 3.8.1 in Report #3002025583: IEEE 1547-2018 OpenDER Model
+    EPRI Report Reference: Section 3.8.1 in Report #3002026631: IEEE 1547-2018 OpenDER Model
     """
 
     def __init__(self, der_obj):